# api/openapi.yaml
openapi: 3.1.0
info:
  title: E-Commerce Backend API
  version: "1.0.0"
  description: >
    Shared REST contract implemented identically by three independent backends
    (Rust/axum, Go/chi, Python/FastAPI). All money values are integer **cents**.
    Errors use RFC 7807 `application/problem+json`. Protected endpoints require
    `Authorization: Bearer <access_token>`.
servers:
  - url: http://localhost:8080
    description: Local (any of the three implementations)

tags:
  - name: auth
  - name: catalog
  - name: cart
  - name: orders
  - name: ops

paths:
  /v1/auth/register:
    post:
      tags: [auth]
      summary: Register a new account
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RegisterRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/RegisterResponse" }
        "409": { $ref: "#/components/responses/Problem" }

  /v1/auth/login:
    post:
      tags: [auth]
      summary: Exchange credentials for tokens
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LoginRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401": { $ref: "#/components/responses/Problem" }

  /v1/auth/refresh:
    post:
      tags: [auth]
      summary: Rotate the refresh token (reuse revokes the family)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/TokenResponse" }
        "401": { $ref: "#/components/responses/Problem" }

  /v1/auth/logout:
    post:
      tags: [auth]
      summary: Revoke a refresh-token family (idempotent)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refresh_token]
              properties:
                refresh_token: { type: string }
      responses:
        "204": { description: No Content }

  /v1/auth/me:
    get:
      tags: [auth]
      summary: Current user profile
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/UserResponse" }
        "401": { $ref: "#/components/responses/Problem" }

  /v1/products:
    get:
      tags: [catalog]
      summary: List products (keyset pagination, search, filter)
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
        - { name: category_id, in: query, schema: { type: string, format: uuid } }
        - { name: q, in: query, schema: { type: string }, description: Full-text search term }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ProductPage" }
    post:
      tags: [catalog]
      summary: Create a product (requires catalog:write)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProductRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "403": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }

  /v1/products/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [catalog]
      summary: Get a product
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/Problem" }
    patch:
      tags: [catalog]
      summary: Update a product (requires catalog:write)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateProductRequest" }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Product" }
        "404": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [catalog]
      summary: Archive a product (soft delete, requires catalog:write)
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: No Content }
        "404": { $ref: "#/components/responses/Problem" }

  /v1/categories:
    get:
      tags: [catalog]
      summary: List categories
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
    post:
      tags: [catalog]
      summary: Create a category (requires catalog:write)
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateCategoryRequest" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }

  /v1/cart:
    get:
      tags: [cart]
      summary: Get the current user's cart
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
    delete:
      tags: [cart]
      summary: Clear the cart
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: No Content }

  /v1/cart/items:
    post:
      tags: [cart]
      summary: Add an item to the cart
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AddItemRequest" }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "404": { $ref: "#/components/responses/Problem" }

  /v1/cart/items/{product_id}:
    parameters:
      - { name: product_id, in: path, required: true, schema: { type: string, format: uuid } }
    patch:
      tags: [cart]
      summary: Set the quantity of a cart line
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [quantity]
              properties:
                quantity: { type: integer, minimum: 1, maximum: 1000 }
      responses:
        "200":
          description: Updated cart
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Cart" }
        "404": { $ref: "#/components/responses/Problem" }
    delete:
      tags: [cart]
      summary: Remove a cart line
      security: [{ bearerAuth: [] }]
      responses:
        "204": { description: No Content }

  /v1/orders:
    get:
      tags: [orders]
      summary: List the current user's orders (keyset pagination)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 20, maximum: 100 } }
        - { name: cursor, in: query, schema: { type: string } }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/OrderPage" }
    post:
      tags: [orders]
      summary: Checkout the cart into an order (ACID, no oversell)
      security: [{ bearerAuth: [] }]
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "400": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }

  /v1/orders/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    get:
      tags: [orders]
      summary: Get an order
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/Problem" }

  /v1/orders/{id}/cancel:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
    post:
      tags: [orders]
      summary: Cancel an order and restock items
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Order" }
        "404": { $ref: "#/components/responses/Problem" }
        "409": { $ref: "#/components/responses/Problem" }

  /healthz:
    get: { tags: [ops], summary: Liveness, responses: { "200": { description: ok } } }
  /readyz:
    get: { tags: [ops], summary: Readiness (checks DB), responses: { "200": { description: ready }, "503": { description: degraded } } }
  /metrics:
    get: { tags: [ops], summary: Prometheus metrics, responses: { "200": { description: metrics } } }

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  responses:
    Problem:
      description: RFC 7807 problem document
      content:
        application/problem+json:
          schema: { $ref: "#/components/schemas/Problem" }

  schemas:
    Problem:
      type: object
      properties:
        type: { type: string }
        title: { type: string }
        status: { type: integer }
        detail: { type: string }
        errors: {}

    RegisterRequest:
      type: object
      required: [email, password, full_name]
      properties:
        email: { type: string, format: email }
        password: { type: string, minLength: 8, maxLength: 128 }
        full_name: { type: string, minLength: 1, maxLength: 200 }

    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: { type: string, format: email }
        password: { type: string }

    TokenResponse:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        token_type: { type: string, example: Bearer }
        expires_in: { type: integer, description: Seconds until access token expiry }

    UserResponse:
      type: object
      properties:
        id: { type: string, format: uuid }
        email: { type: string }
        full_name: { type: string }
        email_verified: { type: boolean }
        status: { type: string }
        roles: { type: array, items: { type: string } }
        created_at: { type: string, format: date-time }

    RegisterResponse:
      type: object
      properties:
        user: { $ref: "#/components/schemas/UserResponse" }
        tokens: { $ref: "#/components/schemas/TokenResponse" }

    Product:
      type: object
      properties:
        id: { type: string, format: uuid }
        sku: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        price_cents: { type: integer }
        currency: { type: string }
        stock_quantity: { type: integer }
        category_id: { type: [string, "null"], format: uuid }
        status: { type: string, enum: [active, archived] }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }

    CreateProductRequest:
      type: object
      required: [sku, name, slug, price_cents]
      properties:
        sku: { type: string }
        name: { type: string }
        slug: { type: string }
        description: { type: string }
        price_cents: { type: integer, minimum: 0 }
        currency: { type: string, default: USD }
        stock_quantity: { type: integer, minimum: 0, default: 0 }
        category_id: { type: string, format: uuid }

    UpdateProductRequest:
      type: object
      properties:
        name: { type: string }
        description: { type: string }
        price_cents: { type: integer, minimum: 0 }
        stock_quantity: { type: integer, minimum: 0 }
        category_id: { type: string, format: uuid }
        status: { type: string, enum: [active, archived] }

    Category:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        slug: { type: string }
        parent_id: { type: [string, "null"], format: uuid }
        created_at: { type: string, format: date-time }

    CreateCategoryRequest:
      type: object
      required: [name, slug]
      properties:
        name: { type: string }
        slug: { type: string }
        parent_id: { type: string, format: uuid }

    ProductPage:
      type: object
      properties:
        items: { type: array, items: { $ref: "#/components/schemas/Product" } }
        next_cursor: { type: [string, "null"] }
        has_more: { type: boolean }

    AddItemRequest:
      type: object
      required: [product_id, quantity]
      properties:
        product_id: { type: string, format: uuid }
        quantity: { type: integer, minimum: 1, maximum: 1000 }

    Cart:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              product_id: { type: string, format: uuid }
              name: { type: string }
              unit_price_cents: { type: integer }
              quantity: { type: integer }
              line_total_cents: { type: integer }
        total_cents: { type: integer }
        currency: { type: string }

    Order:
      type: object
      properties:
        id: { type: string, format: uuid }
        status: { type: string, enum: [pending, confirmed, cancelled] }
        total_cents: { type: integer }
        currency: { type: string }
        created_at: { type: string, format: date-time }
        items:
          type: array
          items:
            type: object
            properties:
              product_id: { type: string, format: uuid }
              product_name: { type: string }
              unit_price_cents: { type: integer }
              quantity: { type: integer }
              line_total_cents: { type: integer }

    OrderPage:
      type: object
      properties:
        items:
          type: array
          items:
            type: object
            properties:
              id: { type: string, format: uuid }
              status: { type: string }
              total_cents: { type: integer }
              currency: { type: string }
              created_at: { type: string, format: date-time }
        next_cursor: { type: [string, "null"] }
        has_more: { type: boolean }